refac(hal): do away with the Architecture trait (#60)
The `Architecture` trait means that pretty much all downstream code that
consumes `hal-core` has to be parameterized over an architecture type,
which is kind of painful. `Architecture` pretty much only provides two
things: definitions of the `PAddr` and `VAddr` types, and an API
contract for architecture implementations.

On most architectures that we are likely to want to target, it's pretty
reasonable to represent `PAddr`s and `VAddr`s as words instead, using
`usize` as the inner repr. This means that for most architectures, HAL
implementations can still be written in separate crates without changing
`hal-core`. For architectures where we _do_ wish to add
architecture-specific logic in the address types, such as validating
that x86 addresses are in canonical form, or which require `PAddr`s or
`VAddr`s to have non-`usize` reprss, we can add `#[cfg(target_arch =
...)]` versions of the address types in `hal-core`.

As far as providing a contract for the shape of the top-level arch APIs,
we can probably solve this through documentation, if there are ever
third-party HAL impls. This may also not matter as much as I thought it
might, since non-arch crates written against the HAL will still probably
be generic over `hal-core` traits for whatever specific HAL APIs they
depend on.

I've also moved the interrupt handler implementations out of the 
`hal-x86_64` crate and into the kernel proper!
