+++ hal-core/src/interrupt/ctx.rs
@@ -1,24 +1,23 @@
-use crate::Architecture;
+use crate::VAddr;
 use core::fmt;
 
 pub trait Context {
-    type Arch: Architecture;
     // TODO(eliza): Registers trait
-    type Registers: fmt::Debug;
+    type Registers: fmt::Debug + fmt::Display;
 
     fn registers(&self) -> &Self::Registers;
     unsafe fn registers_mut(&mut self) -> &mut Self::Registers;
 }
 
 pub trait PageFault: Context {
-    fn fault_vaddr(&self) -> <<Self as Context>::Arch as Architecture>::VAddr;
+    fn fault_vaddr(&self) -> VAddr;
 
     // TODO(eliza): more
 }
 
 pub trait CodeFault: Context {
     fn is_user_mode(&self) -> bool;
-    fn instruction_ptr(&self) -> <<Self as Context>::Arch as Architecture>::VAddr;
+    fn instruction_ptr(&self) -> VAddr;
 }
 
 #[non_exhaustive]
