+++ src/arch/mod.rs
@@ -2,4 +2,4 @@
 pub(crate) mod x86_64;
 
 #[cfg(target_arch = "x86_64")]
-pub(crate) use self::x86_64::*;
+pub use self::x86_64::*;
